Rename business to logic
