+++ features/dashboard/logic/src/main/AndroidManifest.xml
