+++ features/dashboard/logic/src/main/kotlin/com/boswelja/truemanager/dashboard/logic/ExtractSystemInformationData.kt
@@ -1,4 +1,4 @@
-package com.boswelja.truemanager.dashboard.business
+package com.boswelja.truemanager.dashboard.logic
 
 import com.boswelja.truemanager.core.api.v2.system.SystemInfo
 import kotlinx.datetime.Instant
