+++ features/dashboard/logic/src/main/kotlin/com/boswelja/truemanager/dashboard/logic/GetDashboardData.kt
@@ -1,9 +1,9 @@
-package com.boswelja.truemanager.dashboard.business
+package com.boswelja.truemanager.dashboard.logic
 
 import com.boswelja.capacity.Capacity
 import com.boswelja.capacity.CapacityUnit
 import com.boswelja.truemanager.core.api.v2.system.SystemV2Api
-import com.boswelja.truemanager.dashboard.business.DashboardData.NetworkUsageData.AdapterData
+import com.boswelja.truemanager.dashboard.logic.DashboardData.NetworkUsageData.AdapterData
 import com.boswelja.truemanager.data.configuration.DashboardConfiguration
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.delay
