+++ settings.gradle.kts
@@ -27,8 +27,8 @@ include(
     ":core:capacity",
     ":features:apps",
     ":features:auth",
-    ":features:dashboard:business",
     ":features:dashboard:data",
+    ":features:dashboard:logic",
     ":features:dashboard:ui",
     ":features:reporting",
     ":features:storage",
